Fix Vulkan rounded rect borders

+++ tests/AnimationTests.cpp
@@ -40,7 +40,7 @@ public:
 
   std::uint32_t resolveFontId(std::string_view, float, bool) override { return 0; }
 
-  std::vector<std::uint8_t> rasterizeGlyph(std::uint32_t, std::uint16_t, float,
+  std::vector<std::uint8_t> rasterizeGlyph(std::uint32_t, std::uint32_t, float,
                                            std::uint32_t& outWidth,
                                            std::uint32_t& outHeight,
                                            Point& outBearing) override {
